[IMP] hr_holidays: display correct calendar attendances
Before this commit, the attendance and non-working days were
always based to the current calendar of the user's employee.
This is not accurate since we can see other's calendars and
contract changes can lead to calendar changes (from full-time to
part-time,...).
This commit base that view on the viewed employee contracts so that
even a contract change is correctly viewed.

task-3060722

closes odoo/odoo#121153

Related: odoo/enterprise#42125
Signed-off-by: Kevin Baptiste <kba@odoo.com>